refactor: Prepare sort package structure for introducing different sort mechanisms
