+++ internal/topologysort/topologysort.go
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package sort
+package topologysort
 
 import "fmt"
