+++ pkg/project/v2/sort/topologysort/topologysort.go
@@ -1,23 +1,26 @@
-// @license
-// Copyright 2021 Dynatrace LLC
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-// http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
+/*
+ * @license
+ * Copyright 2023 Dynatrace LLC
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
 
 package topologysort
 
 import (
-	"fmt"
 	"github.com/dynatrace/dynatrace-configuration-as-code/v2/internal/log"
-	"github.com/dynatrace/dynatrace-configuration-as-code/v2/internal/sort"
+	"github.com/dynatrace/dynatrace-configuration-as-code/v2/internal/topologysort"
+	"github.com/dynatrace/dynatrace-configuration-as-code/v2/pkg/config/parameter"
+	errors2 "github.com/dynatrace/dynatrace-configuration-as-code/v2/pkg/project/v2/sort/errors"
 	"strings"
 	"sync"
 
@@ -26,109 +29,14 @@ import (
 	"github.com/dynatrace/dynatrace-configuration-as-code/v2/pkg/config"
 	"github.com/dynatrace/dynatrace-configuration-as-code/v2/pkg/config/coordinate"
 	"github.com/dynatrace/dynatrace-configuration-as-code/v2/pkg/config/errors"
-	"github.com/dynatrace/dynatrace-configuration-as-code/v2/pkg/config/parameter"
 	project "github.com/dynatrace/dynatrace-configuration-as-code/v2/pkg/project/v2"
 )
 
-type ProjectsPerEnvironment map[string][]project.Project
+type projectsPerEnvironment map[string][]project.Project
 
-type CircularDependencyParameterSortError struct {
-	Location           coordinate.Coordinate          `json:"location"`
-	EnvironmentDetails errors.EnvironmentDetails      `json:"environmentDetails"`
-	ParameterName      string                         `json:"parameterName"`
-	DependsOn          []parameter.ParameterReference `json:"dependsOn"`
-}
-
-func (e CircularDependencyParameterSortError) Coordinates() coordinate.Coordinate {
-	return e.Location
-}
-
-func (e CircularDependencyParameterSortError) LocationDetails() errors.EnvironmentDetails {
-	return e.EnvironmentDetails
-}
-
-var (
-	_ errors.DetailedConfigError = (*CircularDependencyParameterSortError)(nil)
-)
-
-func (e CircularDependencyParameterSortError) Error() string {
-	return fmt.Sprintf("%s: circular dependency detected. check parameter dependencies: %s",
-		e.ParameterName, joinParameterReferencesToString(e.DependsOn))
-}
-
-func joinParameterReferencesToString(refs []parameter.ParameterReference) string {
-	switch len(refs) {
-	case 0:
-		return ""
-	case 1:
-		return refs[0].String()
-	}
-
-	result := strings.Builder{}
-
-	for _, ref := range refs {
-		result.WriteString(ref.String())
-		result.WriteString(", ")
-	}
-
-	return result.String()
-}
-
-type CircualDependencyProjectSortError struct {
-	Environment string `json:"environment"`
-	Project     string `json:"project"`
-	// slice of project ids
-	DependsOnProjects []string `json:"dependsOnProjects"`
-}
-
-func (e CircualDependencyProjectSortError) Error() string {
-	return fmt.Sprintf("%s:%s: circular dependency detected.\n check project dependencies: %s",
-		e.Environment, e.Project, strings.Join(e.DependsOnProjects, ", "))
-}
-
-type CircularDependencyConfigSortError struct {
-	Location    coordinate.Coordinate   `json:"location"`
-	Environment string                  `json:"environment"`
-	DependsOn   []coordinate.Coordinate `json:"dependsOn"`
-}
-
-func (e CircularDependencyConfigSortError) Error() string {
-	return fmt.Sprintf("%s:%s: is part of circular dependency.\n depends on: %s",
-		e.Environment, e.Location, joinCoordinatesToString(e.DependsOn))
-}
-
-func joinCoordinatesToString(coordinates []coordinate.Coordinate) string {
-	switch len(coordinates) {
-	case 0:
-		return ""
-	case 1:
-		return coordinates[0].String()
-	}
-
-	result := strings.Builder{}
-
-	for _, c := range coordinates {
-		result.WriteString(c.String())
-		result.WriteString(", ")
-	}
-
-	return result.String()
-}
-
-var (
-	_ error = (*CircularDependencyConfigSortError)(nil)
-	_ error = (*CircualDependencyProjectSortError)(nil)
-	_ error = (*CircularDependencyParameterSortError)(nil)
-)
-
-type ParameterWithName struct {
-	Name      string
-	Parameter parameter.Parameter
-}
-
-func (p *ParameterWithName) IsReferencing(config coordinate.Coordinate, param ParameterWithName) bool {
-	for _, ref := range p.Parameter.GetReferences() {
-		if ref.Config == config && ref.Property == param.Name {
+func parameterReference(sourceParam parameter.NamedParameter, config coordinate.Coordinate, targetParam parameter.NamedParameter) bool {
+	for _, ref := range sourceParam.Parameter.GetReferences() {
+		if ref.Config == config && ref.Property == targetParam.Name {
 			return true
 		}
 	}
@@ -136,11 +44,11 @@ func (p *ParameterWithName) IsReferencing(config coordinate.Coordinate, param Pa
 	return false
 }
 
-func SortParameters(group string, environment string, conf coordinate.Coordinate, parameters config.Parameters) ([]ParameterWithName, []error) {
-	parametersWithName := make([]ParameterWithName, 0, len(parameters))
+func SortParameters(group string, environment string, conf coordinate.Coordinate, parameters config.Parameters) ([]parameter.NamedParameter, []error) {
+	parametersWithName := make([]parameter.NamedParameter, 0, len(parameters))
 
 	for name, param := range parameters {
-		parametersWithName = append(parametersWithName, ParameterWithName{
+		parametersWithName = append(parametersWithName, parameter.NamedParameter{
 			Name:      name,
 			Parameter: param,
 		})
@@ -151,14 +59,14 @@ func SortParameters(group string, environment string, conf coordinate.Coordinate
 	})
 
 	matrix, inDegrees := parametersToSortData(conf, parametersWithName)
-	sorted, sortErrs := sort.TopologySort(matrix, inDegrees)
+	sorted, sortErrs := topologysort.TopologySort(matrix, inDegrees)
 
 	if len(sortErrs) > 0 {
 		errs := make([]error, len(sortErrs))
 		for i, sortErr := range sortErrs {
 			param := parametersWithName[sortErr.OnId]
 
-			errs[i] = &CircularDependencyParameterSortError{
+			errs[i] = &errors2.CircularDependencyParameterSortError{
 				Location: conf,
 				EnvironmentDetails: errors.EnvironmentDetails{
 					Group:       group,
@@ -172,7 +80,7 @@ func SortParameters(group string, environment string, conf coordinate.Coordinate
 
 	}
 
-	result := make([]ParameterWithName, 0, len(parametersWithName))
+	result := make([]parameter.NamedParameter, 0, len(parametersWithName))
 
 	for i := len(sorted) - 1; i >= 0; i-- {
 		result = append(result, parametersWithName[sorted[i]])
@@ -181,7 +89,7 @@ func SortParameters(group string, environment string, conf coordinate.Coordinate
 	return result, nil
 }
 
-func parametersToSortData(conf coordinate.Coordinate, parameters []ParameterWithName) ([][]bool, []int) {
+func parametersToSortData(conf coordinate.Coordinate, parameters []parameter.NamedParameter) ([][]bool, []int) {
 	numParameters := len(parameters)
 	matrix := make([][]bool, numParameters)
 	inDegrees := make([]int, numParameters)
@@ -194,7 +102,7 @@ func parametersToSortData(conf coordinate.Coordinate, parameters []ParameterWith
 				continue
 			}
 
-			if p.IsReferencing(conf, param) {
+			if parameterReference(p, conf, param) {
 				logDependency("Config Parameter", p.Name, param.Name)
 				matrix[i][j] = true
 				inDegrees[i]++
@@ -205,7 +113,7 @@ func parametersToSortData(conf coordinate.Coordinate, parameters []ParameterWith
 	return matrix, inDegrees
 }
 
-func GetSortedConfigsForEnvironments(projects []project.Project, environments []string) (map[string][]config.Config, []error) {
+func SortProjects(projects []project.Project, environments []string) (map[string][]config.Config, []error) {
 	sortedProjectsPerEnvironment, errs := sortProjects(projects, environments)
 	if len(errs) > 0 {
 		return nil, errs
@@ -252,7 +160,7 @@ func getConfigs(m map[string][]config.Config) []config.Config {
 func sortConfigs(configs []config.Config) ([]config.Config, []error) {
 	matrix, inDegrees := configsToSortData(configs)
 
-	sorted, sortErrs := sort.TopologySort(matrix, inDegrees)
+	sorted, sortErrs := topologysort.TopologySort(matrix, inDegrees)
 
 	if len(sortErrs) > 0 {
 		return nil, parseConfigSortErrors(sortErrs, configs)
@@ -325,10 +233,10 @@ func configsToSortData(configs []config.Config) ([][]bool, []int) {
 	return matrix, inDegrees
 }
 
-// parseConfigSortErrors turns [sort.TopologySortError] into [CircularDependencyConfigSortError]
+// parseConfigSortErrors turns [topologysort.TopologySortError] into [CircularDependencyConfigSortError]
 // for each config still has an edge to another after sorting an error will be created by aggregating the sort errors
-func parseConfigSortErrors(sortErrs []sort.TopologySortError, configs []config.Config) []error {
-	depErrs := make(map[coordinate.Coordinate]CircularDependencyConfigSortError)
+func parseConfigSortErrors(sortErrs []topologysort.TopologySortError, configs []config.Config) []error {
+	depErrs := make(map[coordinate.Coordinate]errors2.CircularDependencyConfigSortError)
 
 	for _, sortErr := range sortErrs {
 		conf := configs[sortErr.OnId]
@@ -340,7 +248,7 @@ func parseConfigSortErrors(sortErrs []sort.TopologySortError, configs []config.C
 				err.DependsOn = append(err.DependsOn, conf.Coordinate)
 				depErrs[dependingConfig.Coordinate] = err
 			} else {
-				depErrs[dependingConfig.Coordinate] = CircularDependencyConfigSortError{
+				depErrs[dependingConfig.Coordinate] = errors2.CircularDependencyConfigSortError{
 					Location:    dependingConfig.Coordinate,
 					Environment: dependingConfig.Environment,
 					DependsOn:   []coordinate.Coordinate{conf.Coordinate},
@@ -359,21 +267,21 @@ func parseConfigSortErrors(sortErrs []sort.TopologySortError, configs []config.C
 	return errs
 }
 
-func sortProjects(projects []project.Project, environments []string) (ProjectsPerEnvironment, []error) {
+func sortProjects(projects []project.Project, environments []string) (projectsPerEnvironment, []error) {
 	var errs []error
 
-	resultByEnvironment := make(ProjectsPerEnvironment)
+	resultByEnvironment := make(projectsPerEnvironment)
 
 	for _, env := range environments {
 		matrix, inDegrees := projectsToSortData(projects, env)
 
-		sorted, sortErrs := sort.TopologySort(matrix, inDegrees)
+		sorted, sortErrs := topologysort.TopologySort(matrix, inDegrees)
 
 		if len(sortErrs) > 0 {
 			for _, sortErr := range sortErrs {
 				p := projects[sortErr.OnId]
 
-				errs = append(errs, &CircualDependencyProjectSortError{
+				errs = append(errs, &errors2.CircualDependencyProjectSortError{
 					Environment:       env,
 					Project:           p.Id,
 					DependsOnProjects: p.Dependencies[env],
